Split Telegram filebox command helpers
